[BugFix] Prevent store add or replace events to save a reaction inside the core data
Now, it prevents to store add/replace event if the message was not inside the store. This will prevent to get incorrect message on loading a reactions after add/replace. For example, after adding a reaction, once opening the thread it will show cached version of reactions with one reaction in it only.